remove event database set
